+++ packages/components/src/components/Image/Image.stories.ts
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import { constants } from '@/utils/constants';
+import { placeholderImage } from '@/utils/constants';
 import { Image, ImageComponentDefinition } from '.';
 
 // More on how to set up stories at: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
@@ -23,7 +23,7 @@ type Story = StoryObj<typeof meta>;
 // More on writing stories with args: https://storybook.js.org/docs/react/writing-stories/args
 export const Default: Story = {
   args: {
-    src: constants.placeholderImage,
+    src: placeholderImage,
     width: 300,
   },
 };