+++ packages/core/src/constants.ts
@@ -166,7 +166,7 @@ export const CF_STYLE_ATTRIBUTES = [
   'cfBackgroundImageAlignmentHorizontal',
 ];
 
-export const EMPTY_CONTAINER_HEIGHT = '120px';
+export const EMPTY_CONTAINER_HEIGHT = '80px';
 
 export enum PostMessageMethods {
   REQUEST_ENTITIES = 'REQUEST_ENTITIES',