+++ packages/visual-editor/src/store/zone.ts
@@ -9,45 +9,27 @@ export interface Zone {
 
 interface State {
   hoveringZone: string;
-  hoveringSection: string;
-  sectionsWithZones: Record<string, boolean>;
   zones: Record<string, Zone>;
 }
 
 interface Actions {
   upsertZone: (id: string, data: Partial<Zone>) => void;
   setHoveringZone: (zoneId: string) => void;
-  setHoveringSection: (sectionId: string) => void;
-  addSectionWithZone: (sectionId: string) => void;
 }
 
 export const useZoneStore = create<Actions & State>()((set) => ({
   zones: {},
-  hoveringSection: '',
   hoveringZone: '',
-  sectionsWithZones: {},
   setHoveringZone(zoneId) {
     set({
       hoveringZone: zoneId,
     });
   },
-  setHoveringSection(sectionId) {
-    set({
-      hoveringSection: sectionId,
-    });
-  },
-  addSectionWithZone(sectionId) {
-    set(
-      produce((state) => {
-        state.sectionsWithZones[sectionId] = true;
-      })
-    );
-  },
   upsertZone(id, data) {
     set(
       produce((state) => {
         state.zones[id] = { ...(state.zones[id] || {}), ...data };
-      })
+      }),
     );
   },
 }));